feat: Analyze videos
Add functionality to determine basic video information.
This for now only detects video length and if the video has sound.
Duration is stored in the duration attribute.
A new attribute "semantic/video_has_sound" is added to store sound info.

Newly uploaded videos are automatically analyzed.
The settings page in the UI provides a toggle to analyze old videos.