feat(logfs): Store indexes to speed up re-opening and prevent writers from blocking readers
Implement storing of indexes that contain the keyspace.

This tremendously speeds up opening existing logs with large history,
since it prevents having to do a full scan of the entire log.

Indexes are written after configurable log sequence intervals.

Note: a partial index feature that builds on previous indexes is
prepared but not yet fully implemented. Once implemented it reduces the
size required for indexes, because new index entries can build on top of
previous ones.

As a side effect this commit also changes the locking system , which
leads to writers not blocking readers, a significant improvement.