feat: Implement logdb history compaction
Introduce a way to "compact" the history of a LogFs db by rewriting all
entities.

To prevent the need to re-write the whole log, this is done by opening a
"new" db in the log with a new prefix. Then migrations are run, all
entities are copied over, and finally the old event log keys are
deleted, and the new keys are renamed to fit the default prefix.